hlq修改

+++ HotelManagerSystem/src/main/java/com/woniuxy/controller/OrderController.java
@@ -59,7 +59,7 @@ public class OrderController {
         user = new User();
         user.setUser_id(1);
         user.setRole_id(1);
-        if (user.getRole_id() == 1) {
+        if (user.getRole_id() == 3) {
             order.setUser(user);
         }
         if (cxfs.equals("当前预定")) {
@@ -179,11 +179,6 @@ public class OrderController {
     @RequestMapping("/createYdOrder")
     public String createYdOrder(Reserve reserve, HttpServletRequest request,HttpServletResponse response) throws ParseException {
         //根据房间类型、选择数量、入住时间、退房时间查询数据库，获取房间
-    	System.out.println(reserve);
-    	List<House> rooms = houseService.allAvailableTypeRooms(reserve.getReserve_checkintime(), reserve.getReserve_checkouttime(), reserve.getHouseType().getHouse_type_id());
-        System.out.println(rooms.size()+"-----------------");
-        
-        
         List<House> houses = houseService.addOperation(reserve.getReserve_checkintime(),reserve.getReserve_checkouttime()
                 ,reserve.getHouseType().getHouse_type_id(),reserve.getHouse_number());
         
@@ -208,7 +203,7 @@ public class OrderController {
 
 
         //设置订单状态
-        if (user.getRole_id() == 1) {
+        if (user.getRole_id() == 3) {
             order.setOrder_state(0);
             order.setFlag(0);
         } else if (user.getRole_id() == 2) {
@@ -260,17 +255,11 @@ public class OrderController {
         //新增order、item
         boolean flag = orderService.createOrder(order);
         
-        //支付测试
+        //支付测试(押金)
         PayController pc=new PayController();
-        pc.payMoney(response, order.getOrder_number(), order.getOrder_totalpay()+"", order.getOrder_number()+order.getOrder_totalpay(), "");
-   
+        pc.payMoney(response, order.getOrder_number(), order.getOrder_deposit()+"",
+                order.getOrder_number()+order.getOrder_deposit(), "");
 
-        //线上支付，跳转支付页面
-        if (flag && order.getOrder_state() == 0 && order.getFlag() == 0) {
-        	
-        } else {
-
-        }
         return null;
 
     }
@@ -280,27 +269,52 @@ public class OrderController {
      *
      * @return
      */
-    public String qcOrder(Order order) {
+    @RequestMapping("/qcOrder")
+    @ResponseBody
+    public String qcOrder(Order order) throws ParseException {
         //先查询这个订单是否存在，并且为正常预定状态
         order = orderService.queryOrderByOrderNumber(order);
         //存在
         if (order != null && order.getOrder_state() == 1 && order.getFlag() == 1) {
-            //修改订单状态为取消
-            orderService.deleteOrder(order);
-
-            //然后退钱
-
-            //还原房间状态
-
-
-            return "取消预定成功";
-        } else {
-            //提示取消失败
-            return "取消预定失败";
+            List<Order> orders = orderService.showAllOrder(order);
+            if (orders.size()==1){
+                //修改订单状态为取消
+                orderService.deleteOrder(order);
+                //还原房间状态
+                for (Item item:orders.get(0).getItems()
+                ) {
+                    houseService.deleteDateHouseOperation(item.getHouse().getHouse_id(),item.getItem_checkintime()
+                            ,item.getItem_checkouttime());
+                }
+                //然后退钱
+                return "取消预定成功";
+            }
         }
-
+        //提示取消失败
+        return "取消预定失败";
     }
 
+    /**
+     * 结账
+     * 传入订单编号
+     * @param order
+     * @return
+     */
+    @RequestMapping("/payAccounts")
+    @ResponseBody
+    public String payAccounts(Order order,HttpServletResponse response){
+        //根据order_number查询order信息
+        order = orderService.queryOrderByOrderNumber(order);
+        //满足已入住且支付押金，存在
+        if (order!=null&&order.getOrder_state()==2&&order.getFlag()==1){
+            PayController pc=new PayController();
+            pc.payMoney(response, order.getOrder_number(), order.getOrder_totalpay()+"",
+                    order.getOrder_number()+order.getOrder_totalpay(), "");
+            return null;
+        }
+        return "结账失败";
+
+    }
 
     /**
      * 修改预定
